+++ frontend/src/languages/english.js
@@ -1,4 +1,12 @@
 const en = {
+  format: {
+    currencyPositon: "%s %v",
+    numberFormat: {
+      decimal: ".",
+      thousand: ",",
+      precision: 2
+    }
+  },
   common: {
     action: "Action",
     actions: "Actions",
@@ -470,7 +478,8 @@ const en = {
     english: "English",
     french: "French",
     german: "German",
-    portuguese: "Portuguese"
+    portuguese: "Portuguese",
+    georgian: "Gerogian"
   }
 };