+++ frontend/src/reducers.js
@@ -1,6 +1,6 @@
 /**
  * Combine all reducers in this file and export the combined reducers.
- * If we were to do this in store.js, reducers wouldn't be hot reloadable.
+ * If we were doing this in store.js, reducers wouldn't be hot reloadable.
  */
 
 import { connectRouter, LOCATION_CHANGE } from "connected-react-router";